requirements.txt
numpy
scipy
pillow
opencv-python
torch
torchvision
scikit-learn
tensorboardX
